+++ src/NetDoc.Parser/DocumentData/ConstantDocumentData.cs
@@ -1,0 +1,7 @@
+﻿namespace NetDoc.Parser.DocumentData
+{
+    public class ConstantDocumentData : DocumentDataObject
+    {
+        public string Value { get; set; }
+    }
+}
