+++ src/NetDoc.Parser/DocumentData/DocumentDataObject.cs
@@ -1,0 +1,15 @@
+﻿namespace NetDoc.Parser.DocumentData
+{
+    public class DocumentDataObject
+    {
+        public string Name { get; set; }
+
+        public string FullName { get; set; }
+
+        public string DisplayName { get; set; }
+
+        public string Summary { get; set; }
+
+        public string ReturnDescription { get; set; }
+    }
+}
