+++ src/NetDoc.Parser/DocumentData/EventDocumentData.cs
@@ -1,0 +1,6 @@
+﻿namespace NetDoc.Parser.DocumentData
+{
+    public class EventDocumentData : DocumentDataObject
+    {
+    }
+}
