+++ src/NetDoc.Parser/DocumentData/MethodParameterData.cs
@@ -1,0 +1,11 @@
+﻿namespace NetDoc.Parser.DocumentData
+{
+    public class MethodParameterData : DocumentDataObject
+    {
+        public DocumentDataObject Type { get; set; }
+
+        public bool IsOut { get; set; }
+
+        public bool IsRef { get; set; }
+    }
+}
