+++ src/NetDoc.Parser/DocumentData/MethodTypeArgumentData.cs
@@ -1,0 +1,6 @@
+﻿namespace NetDoc.Parser.DocumentData
+{
+    public class MethodTypeArgumentData : DocumentDataObject
+    {
+    }
+}
