+++ src/NetDoc.Parser/DocumentData/NamespaceDocumentData.cs
@@ -1,0 +1,23 @@
+﻿namespace NetDoc.Parser.DocumentData
+{
+    using System.Collections.Generic;
+    using System.Linq;
+
+    public class NamespaceDocumentData : DocumentDataObject
+    {
+        private List<NamedTypeDocumentData> namedTypes = new List<NamedTypeDocumentData>();
+
+        public IEnumerable<NamedTypeDocumentData> NamedTypes
+        {
+            get
+            {
+                return this.namedTypes.OrderBy(m => m.Name).ToArray();
+            }
+        }
+
+        public void AddNamedType(NamedTypeDocumentData data)
+        {
+            this.namedTypes.Add(data);
+        }
+    }
+}
