+++ src/NetDoc.Parser/DocumentData/PropertyDocumentData.cs
@@ -1,0 +1,7 @@
+﻿namespace NetDoc.Parser.DocumentData
+{
+    public class PropertyDocumentData : DocumentDataObject
+    {
+        public DocumentDataObject Type { get; set; }
+    }
+}
