+++ src/NetDoc.Parser/NetDoc.Parser.csproj
@@ -66,7 +66,16 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DocParser.cs" />
-    <Compile Include="DocumentData.cs" />
+    <Compile Include="DocumentData\ConstantDocumentData.cs" />
+    <Compile Include="DocumentData\DocumentData.cs" />
+    <Compile Include="DocumentData\DocumentDataObject.cs" />
+    <Compile Include="DocumentData\EventDocumentData.cs" />
+    <Compile Include="DocumentData\MethodDocumentData.cs" />
+    <Compile Include="DocumentData\MethodParameterData.cs" />
+    <Compile Include="DocumentData\MethodTypeArgumentData.cs" />
+    <Compile Include="DocumentData\NamedTypeDocumentData.cs" />
+    <Compile Include="DocumentData\NamespaceDocumentData.cs" />
+    <Compile Include="DocumentData\PropertyDocumentData.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
